Add reentrant version of functions
The sensors library uses a global config that needs to be initialized
and then cleaned up.

Gather the config into a struct, add a version of the functions that
accept the config to operate on, and use one global config internally to
implement the existing interface.

This is not really reentrant because the parser isn't because the
number parsing isn't because locale isn't, and libc number parsing
cannot be done without locale.

It is sufficient for language runtimes that do user thread switching in
the runtime, and don't interrupt C code. The parser won't be interrupted
in those. In software that uses kernel threads a lock must be held
around the init call.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>