Emit missing sensor warning only if SENSORS_SYSFS_ROOT is defined
The lm-sensors library is widely used in various software, which could
be run in VMs where HWMON support may be absent (e.g. cloud kernels in
Debian). To avoid a confusing "/sys: No sensors at /sys/class/hwmon" on
the console or in the logs, only emit the message if the user specifies
an alternative path through SENSORS_SYSFS_ROOT. This behavior mirrors
what is done for the "Not a sysfs mount" message.

Fixes: 89548dab2a617 ("Add SENSORS_SYSFS_ROOT environment support")